Move fetch_tables_to_postgres to common logic, and add test
